added conmments on other docs
