+++ components/Colors.js
@@ -1,3 +1,4 @@
+// Desc: This file contains all the colors used in the app
 const colors = {
     card: 'darkgrey',
     shadow: 'black',
@@ -6,6 +7,7 @@ const colors = {
     disableConfirm: 'white',
     confirm: 'blue',
     alert: '#B22222',
+    transparent: 'rgba(0, 0, 0, 0)',
 }
 
 export default colors;
